+++ tetra/templatetags/tetra.py
@@ -1,3 +1,5 @@
+import logging
+
 from django import template
 from django.template.loader_tags import BlockNode, BLOCK_CONTEXT_KEY
 from django.apps import apps
@@ -9,6 +11,8 @@ from threading import local
 
 from ..components import ComponentException
 from ..component_register import resolve_component
+
+logger = logging.getLogger(__name__)
 
 
 class TetraTemplateTagException(Exception):
@@ -136,7 +140,7 @@ def do_component(parser, token):
     if "**context" in bits_grouped["context:"]:
         if len(bits_grouped["context:"]) > 1:
             raise template.TemplateSyntaxError(
-                f"Component '{component_name}' multiple context arguments as well a "
+                f"Component '{component_name}' has multiple context arguments as well a "
                 "**context for all context."
             )
         context_args = ALL_CONTEXT
@@ -171,7 +175,7 @@ def do_component(parser, token):
         nodelist = parser.parse((f"/@",))
         if not current_loaded_blocks is None:
             parser.__loaded_blocks = (
-                current_loaded_blocks  # Return origional __loaded_blocks
+                current_loaded_blocks  # Return original __loaded_blocks
             )
         parser.delete_first_token()
 
@@ -242,6 +246,10 @@ class ComponentNode(template.Node):
             )
 
     def render(self, context):
+        """
+        :param context: The template context in which the component is being rendered. It must include the "request" attribute.
+        :return: The rendered component as a tag string or directly rendered component as per the resolved state.
+        """
         Component = resolve_component(context, self.component_name)
         try:
             request = context.request
@@ -254,10 +262,31 @@ class ComponentNode(template.Node):
         resolved_kwargs = {k: v.resolve(context) for k, v in self.kwargs.items()}
         resolved_attrs = {k: v.resolve(context) for k, v in self.attrs.items()}
 
-        if self.context_args == ALL_CONTEXT:
+        extra_context = getattr(Component, "_extra_context", [])
+        if type(extra_context) == str:
+            extra_context = [extra_context]
+
+        if self.context_args == ALL_CONTEXT or "__all__" in extra_context:
             resolved_context = context
         else:
-            resolved_context = template.Context(
+            resolved_context = template.Context()
+            ctx = {}
+            for k in extra_context:
+                if k in context:
+                    ctx[k] = context[k]
+                else:
+                    # this variable is nnot in context!
+                    logger.warning(
+                        f"Component {self} uses '{k}' in _extra_context, "
+                        f"but it is not available in current context."
+                    )
+            if ctx:
+                resolved_context.update(ctx)
+
+        if self.context_args != ALL_CONTEXT:
+            # update context with the explicitly given params. This may not
+            # happen if ALL context is requested directly on the template calling line.
+            resolved_context.update(
                 {k: v.resolve(context) for k, v in self.context_args.items()}
             )